refactor(cli): extract manifest secret domain
